+++ src/main/java/insane96mcp/enhancedai/mixin/WitchMixin.java
@@ -166,7 +166,7 @@ public abstract class WitchMixin extends Raider {
 					}
 				}
 
-				if (!this.hasEffect(MobEffects.INVISIBILITY) && --this.invisibilityCooldown <= 0 && this.getHealth() < this.getMaxHealth() * WitchPotionThrowing.healthThresholdInvisibility) {
+				if (!this.hasEffect(MobEffects.INVISIBILITY) && this.onGround() && --this.invisibilityCooldown <= 0 && this.getHealth() < this.getMaxHealth() * WitchPotionThrowing.healthThresholdInvisibility) {
 					ThrownPotion thrownPotion = new ThrownPotion(this.level(), this);
 					thrownPotion.setItem(MCUtils.setCustomEffects(new ItemStack(Items.SPLASH_POTION), List.of(new MobEffectInstance(MobEffects.INVISIBILITY, 200))));
 					thrownPotion.shoot(0, -1d, 0, 0.1f, 2f);